[6467] * Fixed: possible crash at log output in case sending mail to not existed receiver. Patch provided by ChipLeo. * Fixed: wrong function use for update field value set.

+++ src/game/Player.cpp
@@ -4807,7 +4807,7 @@ void Player::SetSkill(uint32 id, uint16 currVal, uint16 maxVal)
         else                                                //remove
         {
             // clear skill fields
-            SetUInt64Value(PLAYER_SKILL_INDEX(i),0);
+            SetUInt32Value(PLAYER_SKILL_INDEX(i),0);
             SetUInt32Value(PLAYER_SKILL_VALUE_INDEX(i),0);
             SetUInt32Value(PLAYER_SKILL_BONUS_INDEX(i),0);
